Move announce scheduler, because it also announces comments
